moves onNewToken to Message Router Service class
