+++ src/android/FirebaseMessagingRouterService.java
@@ -21,4 +21,16 @@ public class FirebaseMessagingRouterService extends FirebaseMessagingService {
             intercomPushClient.handlePush(getApplication(), remoteMessage.getData());
         }
     }
+
+    @Override
+   public void onNewToken(String token) {
+       // Send a new token to Pushwoosh
+       PushwooshFcmHelper.onTokenRefresh(this, token);
+       sendTokenToAnotherService(token);
+   }
+
+   private void sendTokenToAnotherService(String token) {
+       IntercomPushClient intercomPushClient = new IntercomPushClient();
+       intercomPushClient.sendTokenToIntercom(getApplication(), token);
+   }
 }
